feat: 지원자 수 필드 타입 변경 및 지원 현황 요약 응답 수정
- Recruit.kt의 applicationCount 필드 타입을 Int에서 Long으로 변경
- ApplicationService.kt에서 지원자 수를 계산하는 로직 수정
- ApplicationDto.kt에 지원 현황 요약 응답 추가
- RecruitSearchDto.kt에서 지원 현황 요약 응답 제거
- RecruitService.kt에서 지원자 수 계산 로직 및 지원 현황 요약 응답 로직 수정
- RecruitDto.kt의 applicationCount 필드 타입을 Int에서 Long으로 변경

+++ src/main/kotlin/com/millo/ilhayoung/recruit/domain/Recruit.kt
@@ -27,7 +27,7 @@ data class Recruit(
     val companyContact: String,           // 기업 연락처
     val representativeName: String,       // 대표자 이름
     val status: RecruitStatus = RecruitStatus.ACTIVE, // 공고 상태
-    val applicationCount: Int = 0,        // 지원자 수
+    val applicationCount: Long = 0L,      // 지원자 수
     val viewCount: Long = 0,              // 조회수
     @Version
     val version: Long = 0                 // 버전 관리를 위한 필드
